save nave refactor
