+++ app/containers/Navigation/reducer.js
